% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_teleconnections.R
\name{count_utility_teleconnections}
\alias{count_utility_teleconnections}
\title{count_utility_teleconnections}
\usage{
count_utility_teleconnections(data_dir,
  powerplants_file_path = "water/UCS-EW3-Energy-Water-Database.xlsx",
  utility_file_path = "energy/Electric_Retail_service_Territories/Electric_Retail_service_Territories.shp",
  cities = NULL)
}
\arguments{
\item{data_dir}{root directory for the spatial data ("/pic/projects/im3/teleconnections/data/")}

\item{powerplants_file_path}{path of power plants data file}

\item{utility_file_path}{path of electric retail service areas file}

\item{cities}{a vector of cities to be included in the count. If omitted, all cities will be included.}
}
\description{
count_utility_teleconnections
}
\details{
counts teleconnections for service areas associated with each city
}
