% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_watershed_data.R
\name{count_watershed_data}
\alias{count_watershed_data}
\title{count_watershed_data}
\usage{
count_watershed_data(data_dir, watersheds = NULL)
}
\arguments{
\item{data_dir}{root directory for the spatial data ("/pic/projects/im3/teleconnections/data/")}

\item{watersheds}{a vector of cities to be included in the count. If omitted, all cities will be included.}

\item{watersheds_file_path}{path of watersheds shapefile within data_dir}

\item{powerplants_file_path}{path of power plants data file}

\item{crop_file_path}{path of crop cover raster}

\item{dams_file_path}{path of National Inventory of Dams "NID" point file}

\item{irrigation_file_path}{path of edited demeter irrigation file.}

\item{nuld_file_path}{path of land use raster file.}
}
\description{
count_watershed_data
}
\details{
counts teleconnections associated with water supply catchments associated with each city
}
