% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{get_basins}
\alias{get_basins}
\title{get_basins}
\usage{
get_basins(watersheds_city, huc2_basins)
}
\arguments{
\item{watersheds_city}{city watersheds shape}

\item{huc2_basins}{HUC shapes for the USA}
}
\description{
get_basins
}
\details{
Find all basins that city watersheds overlap with.
}
\author{
Kristian Nelson (kristian.nelson@pnnl.gov)
}
