% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{get_cities}
\alias{get_cities}
\title{get_cities}
\usage{
get_cities()
}
\description{
Read internal data file that specifies all 236 US cities from UWB database.
Provides consistent mapping across cities, intakes and watersheds.
}
\author{
Sean Turner (sean.turner@pnnl.gov)
}
