% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{get_city_utility_mapping}
\alias{get_city_utility_mapping}
\title{get_city_utility_mapping}
\usage{
get_city_utility_mapping()
}
\description{
get_city_utility_mapping
}
\details{
Load plant code file to map cities to Utility IDs
}
\author{
Kristian Nelson (kristian.nelson@pnnl.gov)
}
