% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{get_crop_mapping}
\alias{get_crop_mapping}
\title{get_crop_mapping}
\usage{
get_crop_mapping()
}
\description{
Read internal data file that specifies the GCAM classification for certain crop types.
}
\author{
Kristian Nelson (kristian.nelson@pnnl.gov)
}
