% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{get_demeter_file}
\alias{get_demeter_file}
\title{get_demeter_file}
\usage{
get_demeter_file(irrigation_file_path)
}
\arguments{
\item{irrigation_file_path}{FUll path to the demeter data file}
}
\description{
get_demeter_file
}
\details{
Load demeter file for irrigation and rainfed crop data
}
\author{
Kristian Nelson (kristian.nelson@pnnl.gov)
}
