% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{get_dvsn_to_huc8}
\alias{get_dvsn_to_huc8}
\title{get_dvsn_to_huc8}
\usage{
get_dvsn_to_huc8()
}
\description{
get_dvsn_to_huc8
}
\details{
load dvsn to huc8 table
}
\author{
Kristian Nelson (kristian.nelson@pnnl.gov)
}
