% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{get_endpoints}
\alias{get_endpoints}
\title{get_endpoints}
\usage{
get_endpoints(transfers)
}
\arguments{
\item{method}{retrieve endpoints of all lines in a file}
}
\value{
endpoints of all lines
}
\description{
get_endpoints
}
\author{
Kristian Nelson (kristian.nelson@pnnl.gov)
}
