% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{get_epa_facilities}
\alias{get_epa_facilities}
\title{get_epa_facilities}
\usage{
get_epa_facilities()
}
\description{
get_epa_facilities
}
\details{
import EPA facilities data
}
\author{
Sean Turner (sean.turner@pnnl.gov)
}
