% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_gamut_datsets.R
\name{get_gamut_datasets}
\alias{get_gamut_datasets}
\title{get_gamut_datasets}
\usage{
get_gamut_datasets(download_path, unzip_path)
}
\arguments{
\item{download_path}{path to the downloaded file}

\item{unzip_path}{path to the unzipped file}
}
\description{
get_gamut_datasets
}
\details{
function to download the gamut input datasets
}
