% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{get_hydro_dataset}
\alias{get_hydro_dataset}
\title{get_hydro_dataset}
\usage{
get_hydro_dataset(data_dir, hydro_file_path)
}
\arguments{
\item{data_dir}{your data directory}

\item{hydro_file_path}{hydro plants data file path}
}
\description{
get_hydro_dataset
}
\details{
Load in nlud_id_table.csv and merge with the economic_ids to get class names
}
\author{
Kristian Nelson (kristian.nelson@pnnl.gov)
}
