% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{get_irrigation_bcm}
\alias{get_irrigation_bcm}
\title{get_irrigation_bcm}
\usage{
get_irrigation_bcm()
}
\description{
get_irrigation_bcm
}
\details{
load in irrigation bcm file path
}
\author{
Kristian Nelson (kristian.nelson@pnnl.gov)
}
