% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{get_irrigation_count}
\alias{get_irrigation_count}
\title{get_irrigation_count}
\usage{
get_irrigation_count(irrigation_city)
}
\arguments{
\item{irrigation_city}{irrigation data subsetted by city watersheds}
}
\description{
get_irrigation_count
}
\details{
Count irrigated crops vs rainfed crops and determine irrigation status
}
\author{
Kristian Nelson (kristian.nelson@pnnl.gov)
}
