% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{get_land_category}
\alias{get_land_category}
\title{get_land_category}
\usage{
get_land_category(percent_area)
}
\description{
get_land_category
}
\details{
Classify watershed condition based on percent development and cultivation.
}
\author{
Kristian Nelson (kristian.nelson@pnnl.gov)
}
