% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{get_nlud_names}
\alias{get_nlud_names}
\title{get_nlud_names}
\usage{
get_nlud_names(economic_ids)
}
\arguments{
\item{economic_ids}{dataframe of nlud sector ids}
}
\description{
get_nlud_names
}
\details{
Load in nlud_id_table.csv and merge with the economic_ids to get class names
}
\author{
Kristian Nelson (kristian.nelson@pnnl.gov)
}
