% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{get_raster_nvals}
\alias{get_raster_nvals}
\title{Get raster value count from polygon input areas}
\usage{
get_raster_nvals(raster_file, polygon_shpfile, city_name)
}
\arguments{
\item{raster_file}{character. A full path to the input raster file with file name and extension}

\item{polygon_shpfile}{character. A full path to the input polygon shapefile with file name and extension}

\item{city_name}{character. The target city name in the input polygon shapefile.}
}
\value{
count of unique raster values in target polygons
}
\description{
Get the count of raster values represented in the input raster dataset
when restricted to the input watershed polygons for a target city.
}
\author{
Chris R. Vernon (chris.vernon@pnnl.gov)
}
