% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{get_raster_val_matrix}
\alias{get_raster_val_matrix}
\title{Get raster value matrix from polygon input areas}
\usage{
get_raster_val_matrix(raster_object, polygon)
}
\arguments{
\item{raster_object}{character. An object of class RasterLayer.}

\item{polygon}{character. A polygon to define spatial boundary of raster value counts (e.g. a given city's watersheds)}
}
\value{
matrix of raster values in target polygons
}
\description{
Get the matrix of raster values represented in the input raster dataset
when restricted to the input watershed polygons for a target city.
}
\author{
Chris R. Vernon (chris.vernon@pnnl.gov)
}
