% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{get_runoff_values}
\alias{get_runoff_values}
\title{get_runoff_values}
\usage{
get_runoff_values(
  cropcover_agg,
  runoff_agg,
  lc_values,
  polygon_area,
  land_table
)
}
\arguments{
\item{cropcover_agg}{cropcover raster}

\item{runoff_agg}{runoff raster}

\item{lc_values}{land cover values}

\item{polygon_area}{area of select watershed}

\item{land_table}{land table created by functions}
}
\description{
get_runoff_values
}
\details{
calculate runoff volume in meters cubed per second
}
\author{
Kristian Nelson (kristian.nelson@pnnl.gov)
}
