% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{get_source_contribution}
\alias{get_source_contribution}
\title{get_source_contribution}
\usage{
get_source_contribution(data_dir, file_paths)
}
\description{
get_source_contribution
}
\details{
get a city's water source contribution breakdown
}
\author{
Sean Turner (sean.turner@pnnl.gov)
}
