% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{get_startpoints}
\alias{get_startpoints}
\title{get_startpoints}
\usage{
get_startpoints(transfers)
}
\arguments{
\item{method}{retrieve startpoints of all lines in a file}
}
\value{
start points of all lines
}
\description{
get_startpoints
}
\author{
Kristian Nelson (kristian.nelson@pnnl.gov)
}
