% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{get_teleconnect_table}
\alias{get_teleconnect_table}
\title{get_teleconnect_table}
\usage{
get_teleconnect_table()
}
\description{
get_teleconnect_table
}
\details{
load in teleconnect table
}
\author{
Kristian Nelson (kristian.nelson@pnnl.gov)
}
