% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{get_usgs_flows}
\alias{get_usgs_flows}
\title{get_usgs_flows}
\usage{
get_usgs_flows(DVSN)
}
\description{
get_usgs_flows
}
\details{
get usgs inflow associated with a DVSN
}
\author{
Sean Turner (sean.turner@pnnl.gov)
}
