% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{get_wasteflow_points}
\alias{get_wasteflow_points}
\title{get_wasteflow_points}
\usage{
get_wasteflow_points()
}
\description{
get_wasteflow_points
}
\details{
load in waste flow table and converts to Spatial Point Dataframe
}
\author{
Kristian Nelson (kristian.nelson@pnnl.gov)
}
