% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{get_watershed_usage}
\alias{get_watershed_usage}
\title{get_watershed_usage}
\usage{
get_watershed_usage(city)
}
\arguments{
\item{city}{city that is being analyzed}
}
\description{
get_watershed_usage
}
\details{
load in watershed usage table
}
\author{
Kristian Nelson (kristian.nelson@pnnl.gov)
}
