% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_vis.R
\name{plot_watershed}
\alias{plot_watershed}
\title{plot_watershed}
\usage{
plot_watershed(data_dir,
  watersheds_file_path = "water/CWM_v2_2/World_Watershed8.shp",
  landcover_file_path = "land/NLCD_2016_Land_Cover_L48_20190424/NLCD_2016_Land_Cover_L48_20190424.img",
  crop_file_path = "land/2016_30m_cdls/2016_30m_cdls.img",
  powerplants_file_path = "water/UCS-EW3-Energy-Water-Database.xlsx",
  conus_file_path = "misc/states_21basic/states.shp", city)
}
\arguments{
\item{data_dir}{root directory for the spatial data ("/pic/projects/im3/teleconnections/data/")}

\item{watersheds_file_path}{path of watersheds shapefile within data_dir}

\item{landcover_file_path}{path of NLCD landcover data file}

\item{powerplants_file_path}{path of power plants data file}

\item{city}{character. City to be analysed in the format "City | STATE" (e.g., "Phoenix | AZ")}
}
\description{
plot_watershed
}
\details{
plots the watersheds and relevant features for a chosen city
}
