% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_vis.R
\name{plot_watershed_ggplot}
\alias{plot_watershed_ggplot}
\title{plot_watershed_ggplot}
\usage{
plot_watershed_ggplot(data_dir,
  watersheds_file_path = "water/CWM_v2_2/World_Watershed8.shp",
  crop_file_path = "land/2016_90m_cdls/cdl_lowres_usa.img",
  crop_attributes = "land/2016_90m_cdls/cdl_lowres_usa.img.vat.dbf",
  powerplants_file_path = "water/UCS-EW3-Energy-Water-Database.xlsx",
  conus_file_path = "misc/states_21basic/states.shp",
  dams_file_path = "water/nabd_fish_barriers_2012/nabd_fish_barriers_2012.shp",
  nlud_file_path = "land/usa_nlud_raster.tif",
  citypoint_file_path = "water/CWM_v2_2/City_Centroid.shp", city)
}
\arguments{
\item{data_dir}{root directory for the spatial data ("/pic/projects/im3/teleconnections/data/")}

\item{watersheds_file_path}{path of watersheds shapefile within data_dir}

\item{crop_file_path}{path of CDL img file}

\item{powerplants_file_path}{path of power plants data file}

\item{conus_file_path}{path of CONUS state boundaries shape file}

\item{city}{character. City to be analysed in the format "City | STATE" (e.g., "Phoenix | AZ")}

\item{landcover_file_path}{path of NLCD landcover data file}
}
\description{
plot_watershed_ggplot
}
\details{
plots the watersheds and relevant features for a chosen city
}
