% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{reclassify_raster}
\alias{reclassify_raster}
\title{reclassify_raster}
\usage{
reclassify_raster(crop_cover_levels)
}
\arguments{
\item{crop_cover_levels}{levels of the crop cover raster file.}
}
\description{
reclassify_raster
}
\details{
Read internal data file that specifies the GCAM classification for certain crop types. Reclassify CDL based on GCAM.
}
\author{
Kristian Nelson (kristian.nelson@pnnl.gov)
}
