% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{sup}
\alias{sup}
\title{suppress warnings convenience function}
\usage{
sup(sup)
}
\arguments{
\item{sup}{item you want suppressed}
}
\value{
action
}
\description{
suppress warnings convenience function
}
\author{
Kristian Nelson (kristian.nelson@pnnl.gov)
}
