% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usgs retrieve.R
\name{update_USGS_data}
\alias{update_USGS_data}
\title{update_USGS_data}
\usage{
update_USGS_data(UWB_USGS_Gages_path)
}
\arguments{
\item{UWB_USGS_Gages_path}{full path to WB_USGS_Gages.shp}
}
\description{
update_USGS_data
}
\details{
download latest USGS gage readings to update flows
}
\author{
Sean Turner (sean.turner@pnnl.gov)
}
