% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_watershed_data.R
\name{count_watershed_data}
\alias{count_watershed_data}
\title{count_watershed_data}
\usage{
count_watershed_data(
  data_dir,
  cities = NULL,
  run_all = TRUE,
  file_paths = c(watersheds = "water/CWM_v2_2/World_Watershed8.shp", withdrawal =
    "water/CWM_v2_2/Snapped_Withdrawal_Points.shp", citypoint =
    "water/CWM_v2_2/City_Centroid.shp", powerplants =
    "water/UCS-EW3-Energy-Water-Database.xlsx", crop =
    "land/2016_90m_cdls/cdl_lowres_usa.img", crop_attributes =
    "land/2016_90m_cdls/cdl_lowres_usa.img.vat.dbf", irrigation =
    "land/Version2_USA_Demeter.csv", nlud = "land/usa_nlud_LR.tif", hydro =
    "energy/ORNL_EHAHydroPlant_FY2020revised.xlsx", transfers =
    "water/transfers/USIBTsHUC6_Dickson.shp",      climate =
    "land/kop_climate_classes.tif", HUC4 = "water/USA_HUC4/huc4_to_huc2.shp", population
    = "land/pden2010_60m.tif", runoff = "water/UWSCatch/USA_Mean_Runoff.tif", nhd_flow =
    "water/UWSCatCH/UWSCatCH_Intake_Flows.shp", contributions =
    "water/UWSCatCH/Watershed_Contributions.csv")
)
}
\arguments{
\item{data_dir}{root directory for the spatial data ("/pic/projects/im3/teleconnections/data/")}

\item{cities}{a vector of cities to be included in the count. If omitted, all cities will be included.}

\item{file_paths}{paths to data files}
}
\description{
count_watershed_data
}
\details{
counts teleconnections associated with water supply catchments associated with each city
}
