% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_teleconnections.R
\name{count_watershed_teleconnections}
\alias{count_watershed_teleconnections}
\title{count_watershed_teleconnections}
\usage{
count_watershed_teleconnections(data_dir, cities = NULL)
}
\arguments{
\item{data_dir}{root directory for the spatial data ("/pic/projects/im3/teleconnections/data/")}

\item{cities}{a vector of cities to be included in the count. If omitted, all cities will be included.}
}
\description{
count_watershed_teleconnections
}
\details{
counts teleconnections assoicated with water supply catchments associated with each city
}
