% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_teleconnections.R
\name{count_watershed_teleconnections}
\alias{count_watershed_teleconnections}
\title{count_watershed_teleconnections}
\usage{
count_watershed_teleconnections(data_dir,
  watersheds_file_path = "water/CWM_v2_2/World_Watershed8.shp",
  powerplants_file_path = "water/UCS-EW3-Energy-Water-Database.xlsx",
  crop_file_path = "land/2016_30m_cdls/2016_30m_cdls.img",
  dams_file_path = "water/nabd_fish_barriers_2012/nabd_fish_barriers_2012.shp",
  irrigation_file_path = "land/usa_demeter.csv",
  nlud_file_path = "land/usa_nlud_raster.tif", cities = NULL,
  poly_slices = 40, n_cores = 2)
}
\arguments{
\item{data_dir}{root directory for the spatial data ("/pic/projects/im3/teleconnections/data/")}

\item{watersheds_file_path}{path of watersheds shapefile within data_dir}

\item{powerplants_file_path}{path of power plants data file}

\item{crop_file_path}{path of crop cover raster}

\item{dams_file_path}{path of National Inventory of Dams "NID" point file}

\item{irrigation_file_path}{path of edited demeter irrigation file.}

\item{cities}{a vector of cities to be included in the count. If omitted, all cities will be included.}

\item{poly_slices}{integer for how may parts to split the watersheds polygons into to enable faster zonal stats}

\item{n_cores}{integer for the number of machine cores used to run the polygon slicing function. 2 is default for users with 16GB of RAM.}

\item{nuld_file_path}{path of land use raster file.}
}
\description{
count_watershed_teleconnections
}
\details{
counts teleconnections assoicated with water supply catchments associated with each city
}
