% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_teleconnections.R
\name{count_watershed_teleconnections}
\alias{count_watershed_teleconnections}
\title{count_watershed_teleconnections}
\usage{
count_watershed_teleconnections(data_dir,
  watersheds_file_path = "water/CWM_v2_2/World_Watershed8.shp",
  powerplants_file_path = "water/UCS-EW3-Energy-Water-Database.xlsx",
  crop_file_path = "land/2016_30m_cdls/2016_30m_cdls.img",
  dams_file_path = "water/nabd_fish_barriers_2012/nabd_fish_barriers_2012.shp",
  cities = NULL)
}
\arguments{
\item{data_dir}{root directory for the spatial data ("/pic/projects/im3/teleconnections/data/")}

\item{watersheds_file_path}{path of watersheds shapefile within data_dir}

\item{powerplants_file_path}{path of power plants data file}

\item{crop_file_path}{path of crop cover raster}

\item{dams_file_path}{path of National Inventory of Dams "NID" point file}

\item{cities}{a vector of cities to be included in the count. If omitted, all cities will be included.}
}
\description{
count_watershed_teleconnections
}
\details{
counts teleconnections assoicated with water supply catchments associated with each city
}
