% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{get_runoff_values}
\alias{get_runoff_values}
\title{get_runoff_values}
\usage{
get_runoff_values(
  cropcover_agg,
  runoff_agg,
  lc_values,
  polygon_area,
  land_table
)
}
\description{
get_runoff_values
}
\details{
calculate runoff volume in meters cubed per second
}
\author{
Kristian Nelson (kristian.nelson@pnnl.gov)
}
