% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{get_ucs_power_plants}
\alias{get_ucs_power_plants}
\title{get_ucs_power_plants}
\usage{
get_ucs_power_plants(data_dir,
  ucs_file_path = "water/UCS-EW3-Energy-Water-Database.xlsx",
  method = "sp")
}
\arguments{
\item{data_dir}{root directory for the spatial data ("/pic/projects/im3/teleconnections/data/")}

\item{ucs_file_path}{path of UCS xlsx file within data_dir}

\item{method}{tool used to convert to spatial data; either 'sp' or 'sf'}
}
\description{
Read the UCS database
}
\author{
Sean Turner (sean.turner@pnnl.gov)
}
