% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{suppress_intersect}
\alias{suppress_intersect}
\title{suppress_intersect}
\usage{
suppress_intersect(object1, object2)
}
\arguments{
\item{method}{tool used to intersect 2 objects and remove unnecessary warnings}
}
\value{
intersection of 2 objects
}
\description{
suppress_intersect
}
\author{
Kristian Nelson (kristian.nelson@pnnl.gov)
}
